+++ juntagrico/admins/forms/job_copy_form.py
@@ -2,8 +2,8 @@ import datetime
 
 from django import forms
 from django.contrib import admin
-from django.contrib.admin import widgets
 from django.core.exceptions import ValidationError
+from django.utils.timezone import localtime
 from django.utils.translation import gettext as _
 
 from juntagrico.entity.jobs import RecuringJob
@@ -35,7 +35,7 @@ class JobCopyForm(forms.ModelForm):
 
         self.fields['start_date'].initial = inst.time.date() + \
             datetime.timedelta(days=1)
-        self.fields['time'].initial = inst.time
+        self.fields['time'].initial = localtime(inst.time)
         self.fields['weekdays'].initial = [inst.time.isoweekday()]
 
     def clean(self):