+++ juntagrico/admins/inlines/subscription_type_inlines.py
@@ -3,15 +3,16 @@ from django.core.exceptions import ValidationError
 from django.forms import BaseInlineFormSet
 from django.utils.translation import gettext as _
 
-from juntagrico.entity.subs import Subscription
 from juntagrico.config import Config
+from juntagrico.entity.subs import Subscription
 
 
 class SubscriptionTypeInlineFormset(BaseInlineFormSet):
     def clean(self):
         required_shares = 0
         for form in self.forms:
-            required_shares += form.instance.type.shares
+            if form.instance.pk:
+                required_shares += form.instance.type.shares
         available_shares = sum([member.active_shares_count for member in self.instance._future_members])
         if required_shares > available_shares:
             raise ValidationError(