+++ juntagrico/lifecycle/member.py
@@ -2,7 +2,7 @@ from django.core.exceptions import ValidationError
 from django.utils.translation import gettext as _
 
 from juntagrico.config import Config
-from juntagrico.mailer import AdminNotification
+from juntagrico.mailer import adminnotification
 from juntagrico.signals import member_deactivated, member_created, member_canceled
 
 
@@ -31,13 +31,13 @@ def check_member_consistency(instance):
                 code='invalid')
         if instance.future_subscription is not None and instance.future_subscription.primary_member.pk == instance.pk:
             raise ValidationError(
-                _('Diese/r/s {} ist noch Hauptabonnent in einer/m {}').format(Config.vocabulary('member'), Config.vocabulary('subscription')),
+                _('Diese/r/s {} ist noch HauptbezieherIn in einer/m {}').format(Config.vocabulary('member'), Config.vocabulary('subscription')),
                 code='invalid')
         if instance.subscription is not None and instance.subscription.primary_member.pk == instance.pk:
             raise ValidationError(
-                _('Diese/r/s {} ist noch Hauptabonnent in einer/m {}').format(Config.vocabulary('member'), Config.vocabulary('subscription')),
+                _('Diese/r/s {} ist noch HauptbezieherIn in einer/m {}').format(Config.vocabulary('member'), Config.vocabulary('subscription')),
                 code='invalid')
 
 
 def handle_member_created(sender, instance, **kwargs):
-    AdminNotification.member_created(instance)
+    adminnotification.member_created(instance)