+++ juntagrico/lifecycle/sub.py
@@ -3,7 +3,7 @@ from django.utils import timezone
 from django.utils.translation import gettext as _
 
 from juntagrico.config import Config
-from juntagrico.mailer import AdminNotification
+from juntagrico.mailer import adminnotification
 from juntagrico.signals import sub_activated, sub_deactivated, sub_canceled, sub_created
 from juntagrico.util.lifecycle import handle_activated_deactivated
 
@@ -38,6 +38,8 @@ def handle_sub_deactivated(sender, instance, **kwargs):
     for member in instance.recipients_all_for_state('active'):
         member.old_subscriptions.add(instance)
         member.subscription = None
+        if member.share_set.filter(payback_date__isnull=True).count() == 0:
+            member.inactive = True
         member.save()
 
 
@@ -46,11 +48,18 @@ def handle_sub_canceled(sender, instance, **kwargs):
 
 
 def handle_sub_created(sender, instance, **kwargs):
-    AdminNotification.subscription_created(instance)
+    adminnotification.subscription_created(instance)
 
 
 def check_sub_consistency(instance):
     if instance._old['active'] != instance.active and instance._old['deactivation_date'] is not None:
         raise ValidationError(
             _('Deaktivierte {0} koennen nicht wieder aktiviert werden').format(Config.vocabulary('subscription_pl')),
+            code='invalid')
+    pm_waiting = instance.primary_member in instance.recipients_all_for_state('waiting')
+    pm_active = instance.primary_member in instance.recipients_all_for_state('active')
+    pm_form = instance._future_members and instance.primary_member in instance._future_members
+    if instance.primary_member is not None and not (pm_waiting or pm_active or pm_form):
+        raise ValidationError(
+            _('HauptbezieherIn muss auch {}-BezieherIn sein').format(Config.vocabulary('subscription')),
             code='invalid')