+++ juntagrico/management/commands/create_member_for_superusers.py
@@ -1,6 +1,8 @@
+from django.contrib.auth.models import User
 from django.core.management.base import BaseCommand
+from django.db.models import signals
 
-from juntagrico.models import *
+from juntagrico.entity.member import Member
 
 
 class Command(BaseCommand):