+++ juntagrico/models.py
@@ -1,25 +1,21 @@
+from django.db import models
 from django.db.models import signals
 from django.utils.translation import gettext as _
 
 import juntagrico
-from juntagrico.entity.billing import *
-from juntagrico.entity.extrasubs import *
-from juntagrico.entity.subs import *
-from juntagrico.entity.depot import *
-from juntagrico.entity.member import *
-from juntagrico.entity.share import *
-from juntagrico.entity.jobs import *
-from juntagrico.entity.mailing import *
-from juntagrico.entity.delivery import *
-from juntagrico.entity.listmessage import *
+from juntagrico.entity.extrasubs import ExtraSubscription
+from juntagrico.entity.jobs import Assignment, OneTimeJob, RecuringJob, Job
+from juntagrico.entity.member import Member
+from juntagrico.entity.share import Share
+from juntagrico.entity.subs import Subscription
 from juntagrico.lifecycle.extrasub import extra_sub_pre_save, handle_extra_sub_deactivated, handle_extra_sub_activated
 from juntagrico.lifecycle.job import job_pre_save, handle_job_canceled, handle_job_time_changed
 from juntagrico.lifecycle.member import member_pre_save, member_post_save, handle_member_deactivated, \
     handle_member_created
-from juntagrico.lifecycle.share import share_post_save, handle_share_created
+from juntagrico.lifecycle.share import share_post_save, handle_share_created, share_pre_save
 from juntagrico.lifecycle.sub import sub_pre_save, handle_sub_canceled, handle_sub_deactivated, handle_sub_activated, \
     sub_post_save, handle_sub_created
-from juntagrico.util.signals import register_entities_for_post_init
+from juntagrico.util.signals import register_entities_for_post_init_and_save
 
 
 class SpecialRoles(models.Model):
@@ -60,6 +56,7 @@ signals.pre_save.connect(extra_sub_pre_save, sender=ExtraSubscription)
 juntagrico.signals.extra_sub_activated.connect(handle_extra_sub_activated, sender=ExtraSubscription)
 juntagrico.signals.extra_sub_deactivated.connect(handle_extra_sub_deactivated, sender=ExtraSubscription)
 ''' share handling '''
+signals.pre_save.connect(share_pre_save, sender=Share)
 signals.post_save.connect(share_post_save, sender=Share)
 juntagrico.signals.share_created.connect(handle_share_created, sender=Share)
 ''' member handling '''
@@ -68,4 +65,4 @@ signals.post_save.connect(member_post_save, sender=Member)
 juntagrico.signals.member_created.connect(handle_member_created, sender=Member)
 juntagrico.signals.member_deactivated.connect(handle_member_deactivated, sender=Member)
 ''' lifecycle all post init'''
-register_entities_for_post_init()
+register_entities_for_post_init_and_save()