+++ juntagrico/views.py
@@ -1,26 +1,34 @@
-from datetime import datetime as dt
-
 from django.contrib import auth
 from django.contrib.auth.decorators import login_required
 from django.db.models import Count
 from django.http import HttpResponse
 from django.shortcuts import render, get_object_or_404, redirect
+from django.utils import timezone
 from django.utils.translation import gettext as _
 
+from juntagrico.config import Config
 from juntagrico.dao.activityareadao import ActivityAreaDao
+from juntagrico.dao.assignmentdao import AssignmentDao
 from juntagrico.dao.deliverydao import DeliveryDao
 from juntagrico.dao.depotdao import DepotDao
+from juntagrico.dao.extrasubscriptioncategorydao import ExtraSubscriptionCategoryDao
 from juntagrico.dao.jobdao import JobDao
 from juntagrico.dao.jobextradao import JobExtraDao
 from juntagrico.dao.jobtypedao import JobTypeDao
-from juntagrico.forms import *
-from juntagrico.mailer import FormEmails
-from juntagrico.models import *
+from juntagrico.dao.memberdao import MemberDao
+from juntagrico.entity.depot import Depot
+from juntagrico.entity.jobs import Job, Assignment, ActivityArea
+from juntagrico.entity.member import Member
+from juntagrico.forms import MemberProfileForm, PasswordForm
+from juntagrico.mailer import adminnotification
+from juntagrico.mailer import append_attachements
+from juntagrico.mailer import formemails
+from juntagrico.mailer import membernotification
 from juntagrico.util import addons
 from juntagrico.util.admin import get_job_admin_url
-from juntagrico.util.mailer import append_attachements
-from juntagrico.util.management import *
-from juntagrico.util.messages import *
+from juntagrico.util.management import password_generator, cancel_share
+from juntagrico.util.messages import home_messages, job_messages
+from juntagrico.util.temporal import next_membership_end_date
 
 
 def get_page_dict(request):
@@ -31,7 +39,7 @@ def get_page_dict(request):
 
 def get_menu_dict(request):
     member = request.user.member
-    next_jobs = [a.job for a in AssignmentDao.upcomming_assignments_for_member(member).order_by('job__time')]
+    next_jobs = JobDao.upcomming_jobs_for_member(member)
 
     required_assignments = 0
     if member.subscription is not None:
@@ -56,7 +64,7 @@ def get_menu_dict(request):
     partner_assignments_core = int(
         sum(a.amount for a in partner_assignments if a.is_core()))
     assignmentsrange = list(range(
-        0, max(required_assignments, userassignments_total+partner_assignments_total)))
+        0, max(required_assignments, userassignments_total + partner_assignments_total)))
 
     depot_admin = DepotDao.depots_for_contact(request.user.member)
     area_admin = ActivityAreaDao.areas_by_coordinator(request.user.member)
@@ -80,6 +88,7 @@ def get_menu_dict(request):
         'show_extras': JobExtraDao.all_job_extras().count() > 0,
         'show_deliveries': len(DeliveryDao.deliveries_by_subscription(request.user.member.subscription)) > 0,
         'admin_menus': addons.config.get_admin_menus(),
+        'admin_subscription_menus': addons.config.get_admin_subscription_menu(),
         'user_menus': addons.config.get_user_menus(),
         'messages': [],
 
@@ -121,7 +130,7 @@ def job(request, job_id):
         if Config.assignment_unit() == 'ENTITY':
             amount = job.multiplier
         elif Config.assignment_unit() == 'HOURS':
-            amount = job.multiplier*job.type.duration
+            amount = job.multiplier * job.type.duration
         add = int(num)
         for i in range(add):
             assignment = Assignment.objects.create(
@@ -131,7 +140,7 @@ def job(request, job_id):
                 assignment.job_extras.add(extra)
         assignment.save()
 
-        MemberNotification.job_signup([member.email], job)
+        membernotification.job_signup(member.email, job)
         # redirect to same page such that refresh in the browser or back
         # button does not trigger a resubmission of the form
         return redirect('job', job_id=job_id)
@@ -155,8 +164,7 @@ def job(request, job_id):
             for extra in assignment.job_extras.all():
                 extras.append(extra.extra_type.display_full)
         reachable = member.reachable_by_email is True or request.user.is_staff or job.type.activityarea.coordinator == member
-        participants_summary.append(
-            (name, None, contact_url, reachable, ' '.join(extras)))
+        participants_summary.append((name, contact_url, reachable, ' '.join(extras)))
         emails.append(member.email)
 
     slotrange = list(range(0, job.slots))
@@ -166,7 +174,7 @@ def job(request, job_id):
     job_is_in_past = job.end_time() < timezone.now()
     job_is_running = job.start_time() < timezone.now()
     job_canceled = job.canceled
-    can_subscribe = not (
+    can_subscribe = job.infinite_slots or not (
         job_fully_booked or job_is_in_past or job_is_running or job_canceled)
 
     renderdict = get_menu_dict(request)
@@ -268,7 +276,7 @@ def area_join(request, area_id):
     new_area = get_object_or_404(ActivityArea, id=int(area_id))
     member = request.user.member
     new_area.members.add(member)
-    AdminNotification.member_joined_activityarea(new_area, member)
+    adminnotification.member_joined_activityarea(new_area, member)
     new_area.save()
     return HttpResponse('')
 
@@ -278,7 +286,7 @@ def area_leave(request, area_id):
     old_area = get_object_or_404(ActivityArea, id=int(area_id))
     member = request.user.member
     old_area.members.remove(member)
-    AdminNotification.member_left_activityarea(old_area, member)
+    adminnotification.member_left_activityarea(old_area, member)
     old_area.save()
     return HttpResponse('')
 
@@ -341,7 +349,7 @@ def contact(request):
 
     if request.method == 'POST':
         # send mail to organisation
-        FormEmails.contact(request.POST.get('subject'), request.POST.get('message'), member, request.POST.get('copy'))
+        formemails.contact(request.POST.get('subject'), request.POST.get('message'), member, request.POST.get('copy'))
         is_sent = True
 
     renderdict = get_menu_dict(request)
@@ -364,10 +372,10 @@ def contact_member(request, member_id, job_id):
 
     if request.method == 'POST':
         # send mail to member
-        attachments = []
-        append_attachements(request, attachments)
-        FormEmails.contact_member(request.POST.get('subject'), request.POST.get('message'), member, contact_member,
-                                  request.POST.get('copy'), attachments)
+        files = []
+        append_attachements(request, files)
+        formemails.contact_member(request.POST.get('subject'), request.POST.get('message'), member, contact_member,
+                                  request.POST.get('copy'), files)
         is_sent = True
     job = JobDao.job_by_id(job_id)
     renderdict = get_menu_dict(request)
@@ -426,15 +434,13 @@ def cancel_membership(request):
         member.end_date = end_date
         member.cancelation_date = now
         if member.is_cooperation_member:
-            AdminNotification.member_canceled(member, end_date, message)
+            adminnotification.member_canceled(member, end_date, message)
         else:
             member.inactive = True
 
         member.save()
         for share in member.active_shares:
-            share.cancelled_date = now
-            share.termination_date = end_date
-            share.save()
+            cancel_share(share, now, end_date)
         return redirect('profile')
 
     missing_iban = member.iban == ''
@@ -444,8 +450,8 @@ def cancel_membership(request):
     f_sub = member.future_subscription
     future_active = f_sub is not None and f_sub.state == 'active' and f_sub.state == 'waiting'
     current_active = sub is not None and sub.state == 'active' and sub.state == 'waiting'
-    future = future_active and f_sub.share_overflow-asc < 0
-    current = current_active and sub.share_overflow-asc < 0
+    future = future_active and f_sub.share_overflow - asc < 0
+    current = current_active and sub.share_overflow - asc < 0
     share_error = future or current
     can_cancel = not coop_member or (not missing_iban and not share_error)
 
@@ -462,7 +468,7 @@ def cancel_membership(request):
 
 @login_required
 def send_confirm(request):
-    MemberNotification.email_confirmation(request.user.member)
+    membernotification.email_confirmation(request.user.member)
     renderdict = get_menu_dict(request)
     return render(request, 'info/confirmation_sent.html', renderdict)
 
@@ -502,7 +508,7 @@ def new_password(request):
             pw = password_generator()
             member.user.set_password(pw)
             member.user.save()
-            MemberNotification.reset_password(member.email, pw)
+            membernotification.reset_password(member.email, pw)
 
     renderdict = get_page_dict(request)
     renderdict.update({