+++ docs/release_notes.rst
@@ -101,7 +101,7 @@ Release Notes
 
 
 1.1.9
-----
+-----
  Has migrations
 
 * Added Cookie consent
@@ -114,7 +114,42 @@ Release Notes
 * Links in emails work now also for internal links
 * Old subscriptions are now visible in Member admin
 * Fixed small bug in the size change
-*
+
+
+1.2.0
+-----
+ Has migrations
+
+* Mailer code refactored, new permissions to be notified when a member, subscription or share is created or canceled
+* Jobs are visible if they are on the same day even if they have already started
+* Job time changed bugfix
+* Removed google maps and corresponding setting
+* Members can leave subscription
+* Main member can be changed by Members themselfes
+* Non share holders are deactivated when subscription is deactivated
+* Better gender texts
+* Billing stubs are remoived to extension
+* Fix in upcoming jobs widget
+* Fix in cancelation date calculation
+* Fix in Co Member adding
+
+1.2.1
+-----
+ Has no migrations
+
+* Fix in subscription change view bug from version 1.2.0
+
+1.2.2
+-----
+ Has migrations
+
+* Various date related fixes for subscription change cancellation and job copying
+* Recuring jobs can no have instance specific additional descriptions
+* Extra subscription and types can now be hidden on the depot list
+* The depot overview list is now grouped by day and contains a total
+* Jobs can now have an infinite number of participants
+* Job types can be hidden
+* Fix for the coordinator bug
 
 
 Dev